#96 Fix for norminette
