+++ includes/minishell.h
@@ -155,6 +155,10 @@ int				check_syntax(t_list *token_list);
 int				expand_env_vals(t_expand_state *e_state, t_envs *envs,
 					int exit_status);
 
+t_expand_state	*fill_diff_between_start_curernt_pos(t_expand_state *e_state);
+t_expand_state	*expand_env_vals_core(t_expand_state *e_state, t_envs *envs,
+					int exit_status);
+
 t_list			*check_quote(t_expand_state *e_state);
 
 char			*get_word_in_quote(t_token *token, t_expand_state *e_state, char *word);
